requirements.txt
absl-py==0.15.0
aiohttp==3.8.3
aiosignal==1.3.1
argcomplete==2.0.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
boto==2.49.0
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
crcmod==1.7
cryptography==39.0.0
fasteners==0.18
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Reuploaded==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
flatbuffers==1.12
frozenlist==1.3.3
gast==0.4.0
gcloud==0.18.3
gcs-oauth2-boto-plugin==3.0
google-api-core==2.11.0
google-api-python-client==2.72.0
google-apitools==0.5.32
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-reauth==0.1.1
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
grpcio==1.34.1
gsutil==5.17
gunicorn==20.1.0
h5py==3.1.0
httplib2==0.20.4
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
monotonic==1.6
multidict==6.0.4
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.2.0.32
opt-einsum==3.3.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dotenv==0.21.0
pyu2f==0.1.5
requests==2.28.1
requests-oauthlib==1.3.1
retry-decorator==1.1.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.46
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.12.1
yarl==1.8.2
zipp==3.11.0
